fixing remaining (mostly compilation) issues after initial rebase of cep-15-accord on cep-21-tcm

+++ test/unit/org/apache/cassandra/service/accord/async/AsyncOperationTest.java
@@ -155,7 +155,7 @@ public class AsyncOperationTest
             Assert.assertNull(cfk);
         }));
 
-        int nowInSeconds = FBUtilities.nowInSeconds();
+        long nowInSeconds = FBUtilities.nowInSeconds();
         SinglePartitionReadCommand command = AccordKeyspace.getCommandsForKeyRead(commandStore, key, nowInSeconds);
         try(ReadExecutionController controller = command.executionController();
             FilteredPartitions partitions = FilteredPartitions.filter(command.executeLocally(controller), nowInSeconds))
